+++ src/paperoni/refinement/__init__.py
@@ -1,10 +1,11 @@
-from . import dblp, doi, pubmed
+from . import dblp, doi, pubmed, title
 from .fetch import fetch_all, register_fetch
 
 __all__ = [
     "dblp",
     "doi",
     "pubmed",
+    "title",
     "fetch_all",
     "register_fetch",
 ]
