+++ tests/refinement/test_fetch/test_refine_crossref_doi_10_1109_ICMLC63072_2024_10935257_.yml
@@ -29,7 +29,7 @@ authors:
   display_name: RANWA AL MALLAH
 flags: []
 links:
-- link: 10.1109/ICMLC63072.2024.10935257
+- link: 10.1109/icmlc63072.2024.10935257
   type: doi
 releases:
 - pages: null
