+++ tests/refinement/test_fetch/test_refine_crossref_doi_10_7554_eLife_79919_.yml
@@ -58,7 +58,7 @@ authors:
   display_name: Paul François
 flags: []
 links:
-- link: 10.7554/eLife.79919
+- link: 10.7554/elife.79919
   type: doi
 releases:
 - pages: null
